requirements/main.txt
diff_match_patch
editdistance
ffmpeg
hydra-core
joblib
librosa>=0.10.0 # specify >=0.10.0 so that librosa.get_duration(path=...) will work
numpy>=1.26, <2.0 # module was used numpy 1.x and may crash in 2.x
omegaconf
pandas
rarfile
regex
sox
tabulate
termplotlib
tqdm
gdown
webvtt-py
wget
python-docx
pydub
dask
distributed
jiwer>=3.1.0,<4.0.0
pyarrow>=8.0.0,<14.0.0
datasets>=2.14.0,<3.0.0
h5py
torchcodec
# toloka-kit  # Temporarily disabled due to Toloka's technical pause; keep as reference for past and future API support
# for some processers, additionally https://github.com/NVIDIA/NeMo 2.2.1 is required
# for some processers, additionally nemo_text_processing is required
# for mcv: apt-get update && apt-get upgrade -y && apt-get install -y sox libsox-fmt-all

# for FasterWhisperInference processor is required: 
    # pip install pytorch-lightning nvidia-cublas-cu12 nvidia-cudnn-cu12==9.* faster_whisper
    # export LD_LIBRARY_PATH=`python3 -c 'import os; import nvidia.cublas.lib; import nvidia.cudnn.lib; print(os.path.dirname(nvidia.cublas.lib.__file__) + ":" + os.path.dirname(nvidia.cudnn.lib.__file__))'`
# for vLLMInference processor is required: pip install "optree>=0.13.0" vllm
# for CometoidWMTQualityEstimation processor is required: pip install pymarian
# for FastTextLangIdClassifier processor is required: pip install fasttext
# for ConvertToTarredAudioDatasetConfig processor can be additionally required: pip install lhotse "nemo-toolkit[common]==2.2.1"
